Updated Codes for Animation
Going to pose some questions about how the forcing functions should be done at not at 3 in the morning. Didnt get a chance to try compiling it.  It is not perfect but it should hopefully take not too long to get in compiling shape.  Let me know what the errors are.  I will try and compile it on friday and catch some of the early errors.  X1 and X2 values need to be sent to the main draw function and the states need to be sent.